rename menu.py files to cms_menus.py
resolves
DeprecationWarning: menu.py filename is deprecated, and it will be removed in version 3.4; please rename it to cms_menus.py
