+++ foundation/organisation/cms_app.py
@@ -1,7 +1,7 @@
 from cms.app_base import CMSApp
 from cms.apphook_pool import apphook_pool
 from django.utils.translation import ugettext_lazy as _
-from .menu import ProjectMenu, ThemeMenu
+from .cms_menus import ProjectMenu, ThemeMenu
 
 
 class UnitsAppHook(CMSApp):
