+++ foundation/press/cms_app.py
@@ -3,7 +3,7 @@ from cms.app_base import CMSApp
 from cms.apphook_pool import apphook_pool
 from django.utils.translation import ugettext_lazy as _
 
-from .menu import PressReleaseMenu
+from .cms_menus import PressReleaseMenu
 #  from .menu import PressMentionMenu
